This is the go file
This task consists in training the animals to make a go task.
The rat has to press the lever and keep it pressed for 500 ms, when a light turns on. The rat has 1000 ms to release the lever and get food. If it delays more than that, no food is delivered. The light is turned off when the rat releases the lever.